Rename EsProjectionsClient to ProjectionsClient
